requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cli-helpers==1.1.0
Click==7.0
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.7
django-cors-headers==2.5.0
django-filter==2.1.0
-e git+https://github.com/ehooo/django_mqtt.git@1180e88fe6a9b9bfb41e20d5a76ebba7f7732432#egg=django_mqtt
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
djangorestframework==3.9.2
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.0.1
MarkupSafe==1.1.1
mycli==1.19.0
mysqlclient==1.4.2.post1
paho-mqtt==1.4.0
Pillow==6.0.0
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.3.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1b3
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
tabulate==0.8.3
terminaltables==3.1.0
uritemplate==3.0.0
urllib3==1.24.2
wcwidth==0.1.7
